Network_qse.jl Documentation

Index

Network_qse.AtomicProperties

Network_qse.MultiNewtonRaphson

Network_qse.df_nse_condition

Network_qse.extract_partition_function

Network_qse.initial_guess

Network_qse.logsumexp

Network_qse.nse_condition

Network_qse.prefactor

Network_qse.read_mass_frdm

Network_qse.read_part_frdm

Network_qse.read_species

Network_qse.AtomicProperties — Type

AtomicProperties

Stores all physical properties of an element with charge number Z and atomic number Z. Mass in units of MeV with conversion factor: M * [meverg] / c^2 = [erg]

Network_qse.MultiNewtonRaphson — Method

MultiNewtonRaphson(x::Vector, T, rho, y, ap)

[-7.0010491,-9.1] ∂/∂xᵢ [f₁,...,fₙ] good guess: [-4.394094904641707, -12.915712928215058]

Network_qse.df_nse_condition — Method

df_nse_condition!(J,μ, T,rho, ap)

computes Jacobian ∇f ∈ ℝ²×ℝ² with f ∈ ℝ², μ ∈ ℝ² Contains all 4 partial derivates with respect to the proton and neutron chemical potentials.

Network_qse.extract_partition_function — Method

extract_partition_function()

checks indentical charge and atomic numbers in massfrdm and partfrdm. Interpolates ω(T)

Network_qse.initial_guess — Method

initial_guess(ap_ni56)

index: 438 or 863 or 762 inverse of saha equation with only one species and μₚ = μₙ. returns μ. We assume that the gas is only composed of Nickel 56 and then solve the Saha Equation for the chemical potentials of Protons and neutrons (which we assume as equal).

Network_qse.logsumexp — Method

logsumexp(arr)

max + log(sum(exp{arr - max})) instead of log(sum(exp(arr)))

Network_qse.nse_condition — Method

nse_condition!(res, μ, T, ρ, y, ap; precision=400)

returns zero function, as obtained from condition of mass conservation and charge neutrality for NSE:

log (∑ᵢXᵢ) = 0

log(∑ᵢ(Zᵢ/Aᵢ)Xᵢ / y) = 0

Network_qse.prefactor — Method

prefactor(pf)

returns prefactor of Xi, as given in http://cococubed.asu.edu/codepages/nse.shtml The physical quantities to compute theBoltzmann prefactor for a given element are temperature, density (which are varied), mass and atomic number.

Network_qse.read_mass_frdm — Method

read_mass_frdm()

read ground state energies

Network_qse.read_part_frdm — Method

read_part_frdm()

read in follwing files:

mass-frdm95.dat:

Ground state properties based on the FRDM model Format ––– Each record of the file contains:

Z : charge number A : mass number El : element symbol fl : flag corresponding to 0 if no experimental data available 1 for a mass excess recommended by Audi&Wapstra (1995) 2 for a measured mass from Audi&Wapstra (1995) Mexp : experimental or recommended atomic mass excess in MeV of Audi&Wapstra (1995) Mth : calculated FRDM atomic mass excess in MeV Emic : calculated FRDM microscopic energy in MeV beta2: calculated quadrupole deformation of the nuclear ground-state beta3: calculated octupole deformation of the nuclear ground-state beta4: calculated hexadecapole deformation of the nuclear ground-state beta6: calculated hexacontatetrapole deformation of the nuclear ground-state

The corresponding FORTRAN format is (2i4,1x,a2,1x,i1,3f10.3,4f8.3)

part_frdm.asc:

Each of the two files starts with 4 header lines briefly summarizing the contents of the given columns. This is followed by entries sorted by charge and mass number of the isotope. Each table ends with the line "END OF TABLE". Each entry consists of 5 lines:

Isotope (in standard notation);

Charge number of isotope, mass number of isotope, ground state spin of the isotope;

3-5. Partition functions normalized to the g.s. spin; Third line: Partition functions for the temperatures (in 10^9 K): 0.01, 0.15, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7; Fourth line: Partition functions for the temperatures (in 10^9 K): 0.8, 0.9, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5; Fifth line: Partition functions for the temperatures (in 10^9 K): 4.0, 4.5, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0 . Information for the next isotope starts after the last partition function line.

PIA: ADDED partition functions and A,Z,s for n,p,d,T,3He

Network_qse.read_species — Method

read_species()

reads species from table

Notes

This layout should be customized in docs/src/index.md